feat: ChannelContentController
