+++ src/main/java/com/innovationcamp/messenger/domain/channel/dto/CreateChannelRequestDto.java
@@ -2,14 +2,15 @@ package com.innovationcamp.messenger.domain.channel.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
+import lombok.Setter;
 import lombok.ToString;
 
 @Getter
-@ToString
+@Setter
 @Schema(description = "새 채널 생성 request Dto")
 public class CreateChannelRequestDto {
     private String channelName;
     private String channelPassword;
     private String channelDescription;
-    private Boolean isPrivate = false;
+    private Boolean isPrivate;
 }
