+++ src/main/java/com/innovationcamp/messenger/domain/channel/dto/SignUpChannelRequestDto.java
@@ -1,0 +1,12 @@
+package com.innovationcamp.messenger.domain.channel.dto;
+
+import com.innovationcamp.messenger.domain.channel.entity.UserChannel;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+public class SignUpChannelRequestDto {
+    String userName;
+    String userEmail;
+    String channelPassword;
+}
